Fix thirteen game-core functions wrongly marked __stdcall
Every one of them ends with a plain RET in the retail image, and their
call sites clean the stack themselves -- they are __cdecl like the rest
of the game core.  Marking them __stdcall made the rebuilt bodies emit
`ret <imm>` and dropped the `add esp` from every caller.
ReadPacketSectionData was the same defect and was costing PacketLoad
alone 30 points.

Object files depended only on their .c file, so a header-only edit
never triggered a rebuild and the next comparison silently ran against
a stale object.  Every unit now depends on include/*.h and src/ix/*.h,
which matters here because MSVC 4.1's codegen is sensitive to the
declaration set: adding a prototype can move an operand or a stack slot
in an unrelated file.

PacketLoad also had a dead DecompressPacketSection call after _exit(0),
its switch cases in the wrong order, and an extra local the retail code
does not have.  gametrack was calling WC1's any_enemy_tail instead of
detect_enemy_tail, and wait_for_end_of_music CheckEscaped instead of
WaitForInputKey.  CentreMouseOnCurrentNavObjective drove WC1's cursor
hooks; WC2 just calls SetPersonnelMousePosition.

  PacketLoad                        61.47% -> 100.00%
  CentreMouseOnCurrentNavObjective  88.37% -> 100.00%
  wait_for_end_of_music             (fixed call target, stays 100%)
  gametrack                         68.81% ->  98.08%

Functions at or above 90%: 1380 -> 1384; 100%: 899 -> 904;
average 95.99% -> 96.03%.  Call-target mismatches 145 -> 141.

+++ src/logic.c
@@ -1606,7 +1606,7 @@ void trim_goals(short obj, short amount)
 }
 
 /* Function start: 0x429CAD */
-int report_kilrathi_rout(int mode)
+short report_kilrathi_rout(int mode)
 {
     short obj;
 
@@ -1760,7 +1760,7 @@ short detect_enemy_tail(short obj)
 }
 
 /* Function start: 0x42A1C5 */
-int is_ship_tailing_player_target(short obj)
+short is_ship_tailing_player_target(short obj)
 {
     short target = g_acShipTarget_00495f20[0];
